+++ experimental/css-has-pseudo/.tape.js
@@ -1,0 +1,24 @@
+module.exports = {
+	'basic': {
+		message: 'supports basic usage'
+	},
+	'basic:preserve': {
+		message: 'supports { preserve: false } usage',
+		options: {
+			preserve: false
+		}
+	},
+	'generated-selector-cases': {
+		message: 'correctly handles generated cases',
+		warnings: 1,
+		options: {
+			preserve: false
+		}
+	},
+	'browser': {
+		message: 'prepare CSS for chrome test',
+		options: {
+			preserve: false
+		}
+	},
+}