+++ plugins/postcss-logical/src/lib/transform-float.js
@@ -43,5 +43,7 @@ function rDecl(decl) {
 }
 
 function clean(decl, preserve) {
-	if (!preserve) decl.remove();
+	if (!preserve) {
+		decl.remove();
+	}
 }